requirements.txt
# Steve Morse five-digit decoder (Module A) + SSN Validator (Module B)
playwright>=1.40.0
requests>=2.28.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiogram>=3.4.1
aiohttp>=3.9.0
supabase>=2.5.0
